TST, CLN: Reorganize folder structure
